requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
streamlit
pypdf
langchain_pinecone
python-dotenv
-e .